perf(core): finish the hot-path allocation, cache-age and shared-rule work
Performance:
- DiffStore entries now carry an age, closing the one remaining
  violation of the every-cache-has-an-age rule: a Ready diff is
  recomputed after 5 s (the git-stat TTL), a failure retries after 3 s,
  a Pending entry is never asked twice — and the old answer stays
  published while the refresh is in flight.
- painting borrows: to_line and build_block hand ratatui borrowed
  spans instead of cloning every span's text per frame; the ambiguous-
  width pass walks the buffer's contiguous storage instead of ~10k
  bounds-checked cell_mut calls; convert reads colour names through
  mlua's borrowing accessor in the hottest conversion path.
- the action band resolves, fits and formats its pills once per
  (registry version, width) instead of O(pills × bindings) string work
  per painted frame, and the band settle diff stores a hash of the
  painted cells instead of cloning them.
- the plugin set is indexed once per reload: focusable/floating/slot
  membership/decorators are precomputed lists, slot_mode is read off
  the index instead of a Lua table read per placement query, and the
  resolved arrangement is cached on everything layout.lua can consult
  (size, epoch, store version, chrome rows, reload) — it was a Lua
  call plus a node-by-node conversion per frame.
- the vt100 row extraction is shared per output stamp across the link
  scan, click resolve and OSC 8 repaint (it walked the whole grid per
  reader per frame); Terminals::sync_meta/meta_map split ends the
  per-publish clone of the meta map.
- remote hook provisioning probes dir+file in one round trip; local
  multi-repo spawns create their worktrees concurrently;
  Database::open_existing skips the schema pass for short-lived worker
  reopens inside a process that already ran it.

Structure, in the same stroke:
- session/theme_config splits its 36 palette tables into palettes.rs,
  leaving ThemePreset::palette's dispatch as the visible seam;
  storage/schema splits the forty migration bodies into migrations.rs
  with the step table beside the DDL it migrates toward.
- shell owns the off-local launcher (HostLauncher: ssh/wsl + the sh -c
  quoting divergence) and the PowerShell quoting rule; git, usage and
  notifications delegate instead of carrying the copies that had
  already drifted into three spellings.
- the task/automation CLIs resolve their --session/--repo/--command
  flags through one cli::action::resolve_action.
