+++ src/git/remote.rs
@@ -423,15 +423,7 @@ pub(super) fn run_host_script(host: &HostDef, script: &str, action: &str) -> Res
 ///   shell re-splits, so the `script` must be POSIX-quoted to survive as a
 ///   single `sh -c` argument (mirroring [`git_command`]).
 pub(crate) fn host_shell_c(host: &HostDef, script: &str) -> Command {
-    let mut cmd = host_launcher(host);
-    if host.is_wsl() {
-        cmd.arg("-e").arg("sh").arg("-c").arg(script);
-    } else {
-        cmd.arg(posix_quote("sh"))
-            .arg(posix_quote("-c"))
-            .arg(posix_quote(script));
-    }
-    cmd
+    super::command::launcher_for(host).shell_c(script)
 }
 
 /// Build a `<launcher> powershell -NoProfile -EncodedCommand <base64>`
@@ -478,9 +470,10 @@ pub(super) fn host_probe(host: &HostDef, posix: &str, windows: &str) -> Command 
 
 /// Quote `s` as a PowerShell single-quoted literal. Only `'` is special inside
 /// one (doubled to escape); `\` and `$` are literal, which is what makes it the
-/// right quote for a Windows path.
+/// right quote for a Windows path. The rule lives in
+/// [`crate::shell::powershell_quote`].
 pub(super) fn powershell_quote(s: &str) -> String {
-    format!("'{}'", s.replace('\'', "''"))
+    crate::shell::powershell_quote(s)
 }
 
 /// Write `bytes` to `remote_path` on `host`, creating the parent directory.
@@ -580,70 +573,78 @@ pub fn expand_remote_tilde(host: &HostDef, path: &str) -> Result<String> {
     }
 }
 
-/// Whether `dir` exists as a directory on `host` (a leading `~` expands against
-/// the remote home). `Ok(false)` means the probe *ran* and answered no — a
-/// transport failure is an `Err`, distinguished by the exit code: the script
-/// answers only 0 / [`REMOTE_PROBE_NEGATIVE`] itself.
-pub(crate) fn remote_dir_exists(host: &HostDef, dir: &str) -> Result<bool> {
+/// What one combined probe learned about a guard directory and a file on a
+/// POSIX host — the parsed answer of [`probe_remote_dir_and_file`].
+#[derive(Debug, PartialEq, Eq)]
+pub(crate) enum DirFileProbe {
+    /// The guard directory is absent.
+    NoDir,
+    /// Directory present; the file does not exist.
+    NoFile,
+    /// Directory present; the path exists but is not a regular file.
+    NotFile,
+    /// Directory present; the file exists, with its exact content.
+    File(String),
+}
+
+/// Probe a guard `dir` and read `file` on `host` in **one** round trip (a
+/// leading `~` on either expands against the remote home). Replaces the
+/// dir-probe → file-read pair remote hook provisioning used to make, which
+/// cost two serial ssh connections per `(host, agent)` at spawn time.
+///
+/// A non-zero exit is an `Err` carrying the remote stderr — which covers both
+/// the transport failing and `cat` failing on an existing file (permission
+/// denied); every *answer* travels on stdout as a sentinel line.
+pub(crate) fn probe_remote_dir_and_file(
+    host: &HostDef,
+    dir: &str,
+    file: &str,
+) -> Result<DirFileProbe> {
     let dir = expand_remote_tilde(host, dir)?;
-    let script = format!(
-        "if test -d {}; then exit 0; else exit {REMOTE_PROBE_NEGATIVE}; fi",
-        posix_quote(&dir)
-    );
-    let output = host_shell_c(host, &script)
+    let file = expand_remote_tilde(host, file)?;
+    let output = host_shell_c(host, &dir_file_probe_script(&dir, &file))
         .stderr(Stdio::piped())
         .output()
-        .context("failed to run remote dir probe")?;
-    match output.status.code() {
-        Some(0) => Ok(true),
-        Some(REMOTE_PROBE_NEGATIVE) => Ok(false),
-        _ => {
-            let stderr = reportable_stderr(&output.stderr);
-            anyhow::bail!("remote dir probe failed: {stderr}")
-        }
-    }
+        .context("failed to run remote dir/file probe")?;
+    let stdout = remote_output_or_stderr(output, "dir/file probe")?;
+    parse_dir_file_probe(&String::from_utf8_lossy(&stdout))
 }
 
-/// Read a regular file on `host` (a leading `~` expands against the remote
-/// home). `Ok(None)` = the path doesn't exist; an existing-but-not-regular
-/// path or a transport failure is an `Err` (see [`remote_dir_exists`] for the
-/// exit-code discipline).
-pub(crate) fn read_remote_file(host: &HostDef, path: &str) -> Result<Option<String>> {
-    let path = expand_remote_tilde(host, path)?;
-    let quoted = posix_quote(&path);
-    let script = format!(
-        "if test -f {quoted}; then cat {quoted}; elif test -e {quoted}; then \
-         exit {REMOTE_PROBE_NOT_FILE}; else exit {REMOTE_PROBE_NEGATIVE}; fi"
-    );
-    let output = host_shell_c(host, &script)
-        .stderr(Stdio::piped())
-        .output()
-        .context("failed to run remote file read")?;
-    match output.status.code() {
-        Some(0) => Ok(Some(String::from_utf8_lossy(&output.stdout).into_owned())),
-        Some(REMOTE_PROBE_NEGATIVE) => Ok(None),
-        Some(REMOTE_PROBE_NOT_FILE) => {
-            anyhow::bail!("remote path {path} exists but is not a regular file")
-        }
-        // Anything else includes `cat`'s own failure (e.g. permission denied,
-        // exit 1) — surface its stderr rather than misreporting the file type.
-        _ => {
-            let stderr = reportable_stderr(&output.stderr);
-            anyhow::bail!("remote file read failed: {stderr}")
-        }
-    }
+/// The [`probe_remote_dir_and_file`] shell script (`dir`/`file` already
+/// tilde-expanded, quoted here). Line protocol in the style of
+/// `list_dir_entries_script`'s `!missing`: the first line is one of `@nodir` /
+/// `@notfile` / `@nofile` / `@file`, and everything **after** an `@file` line
+/// is the file's content — which is what makes the sentinels collision-proof:
+/// content that happens to start with `@nodir` sits after the first line and
+/// is never looked at as a sentinel. Pure so the quoting is testable without a
+/// host.
+pub(super) fn dir_file_probe_script(dir: &str, file: &str) -> String {
+    format!(
+        "if ! test -d {dir}; then echo '@nodir'; exit 0; fi; \
+         if test -f {file}; then echo '@file'; cat {file}; \
+         elif test -e {file}; then echo '@notfile'; \
+         else echo '@nofile'; fi",
+        dir = posix_quote(dir),
+        file = posix_quote(file),
+    )
 }
 
-/// Sentinel exit code remote probe scripts use for their "negative" answer, so
-/// it can't be confused with a transport failure (ssh's 255, a launch error,
-/// or the probed command's own 1).
-pub(super) const REMOTE_PROBE_NEGATIVE: i32 = 3;
-
-/// Sentinel for "the path exists but is not a regular file" — distinct from
-/// [`REMOTE_PROBE_NEGATIVE`] *and* from the probed command's own exit 1 (e.g.
-/// `cat` failing on a permission-denied file must not be misreported as a
-/// file-type problem).
-pub(super) const REMOTE_PROBE_NOT_FILE: i32 = 4;
+/// Parse [`dir_file_probe_script`]'s stdout: the sentinel first line, then —
+/// for `@file` — the content verbatim (bytes after the sentinel's newline, so
+/// a missing trailing newline survives). Pure, for testing without a host.
+pub(super) fn parse_dir_file_probe(stdout: &str) -> Result<DirFileProbe> {
+    let (sentinel, content) = match stdout.split_once('\n') {
+        Some((first, rest)) => (first, rest),
+        None => (stdout, ""),
+    };
+    match sentinel {
+        "@nodir" => Ok(DirFileProbe::NoDir),
+        "@nofile" => Ok(DirFileProbe::NoFile),
+        "@notfile" => Ok(DirFileProbe::NotFile),
+        "@file" => Ok(DirFileProbe::File(content.to_string())),
+        other => anyhow::bail!("unexpected dir/file probe answer: {other:?}"),
+    }
+}
 
 /// The longest command line safely below Windows `cmd.exe`'s ~8191-char limit
 /// (the sshd default shell may be `cmd`), leaving headroom for the PowerShell
